LayoutTests/fast/mediastream/MediaStreamConstructor.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<link rel="stylesheet" href="../js/resources/js-test-style.css">
<script src="../js/resources/js-test-pre.js"></script>
</head>
<body>
<p id="description"></p>
<div id="console"></div>
<script>
description("Tests webkitMediaStream.");

var undefined;
var stream;
var newStream;
var nAudio;
var nVideo;

function error() {
    testFailed('Stream generation failed.');
    finishJSTest();
}

function getUserMedia(dictionary, callback) {
    try {
        navigator.webkitGetUserMedia(dictionary, callback, error);
    } catch (e) {
        testFailed('webkitGetUserMedia threw exception :' + e);
        finishJSTest();
    }
}

function gotStream(s) {
    localStream = s;
    testPassed('Got local stream.');
    shouldBe('localStream.getAudioTracks().length', '1');
    shouldBe('localStream.getVideoTracks().length', '1');

    shouldBeTrue("typeof webkitMediaStream === 'function'");

    testConstructor(localStream);

    finishJSTest();
}

function testConstructor(s) {
    stream = s;
    shouldThrow('new webkitMediaStream(document)');
    shouldThrow('new webkitMediaStream([document])');
    shouldThrow('new webkitMediaStream([stream.getAudioTracks()[0], document])');
    shouldThrow('new webkitMediaStream([null])');
    shouldThrow('new webkitMediaStream([undefined])');
    shouldThrow('new webkitMediaStream(null)');
    shouldThrow('new webkitMediaStream(undefined)');

    verifyStream(new webkitMediaStream(), 0, 0);
    verifyStream(new webkitMediaStream([]), 0, 0);

    verifyStream(new webkitMediaStream(s), s.getAudioTracks().length, s.getVideoTracks().length);

    verifyStream(new webkitMediaStream([s.getAudioTracks()[0]]), 1, 0);
    verifyStream(new webkitMediaStream([s.getVideoTracks()[0]]), 0, 1);
    verifyStream(new webkitMediaStream([s.getAudioTracks()[0], s.getVideoTracks()[0]]), 1, 1);
    verifyStream(new webkitMediaStream([s.getVideoTracks()[0], s.getAudioTracks()[0], s.getVideoTracks()[0]]), 1, 1);
}

function verifyStream(s, numAudioTracks, numVideoTracks) {
    newStream = s;
    nAudio = numAudioTracks;
    nVideo = numVideoTracks;

    testPassed('Stream constructed');
    shouldBeNonNull(newStream);
    shouldBeDefined(newStream);
    shouldBe("newStream.constructor.name", "'MediaStream'");
    shouldBe('newStream.getAudioTracks().length', 'nAudio');
    shouldBe('newStream.getVideoTracks().length', 'nVideo');
}

getUserMedia({video:true, audio:true}, gotStream);

window.jsTestIsAsync = true;
window.successfullyParsed = true;
</script>
<script src="../js/resources/js-test-post.js"></script>
</body>
</html>
